COMException addressed
The apparent problems that were happening due to popup concurrency were
not only due to concurrency but also due to COM Exception which is
triggered when excel is suspended during a mouse click. So if a delete
opeariton coincides with a mouse click this exception is triggered. This
exception has been handled in this commit.